Dashboards: Refactor types for dynamic dashboards
