+++ public/app/features/dashboard-scene/edit-pane/ElementSelection.ts
@@ -2,12 +2,9 @@ import { SceneObject, SceneObjectRef, VizPanel } from '@grafana/scenes';
 import { ElementSelectionContextItem } from '@grafana/ui';
 
 import { DashboardScene } from '../scene/DashboardScene';
-import {
-  EditableDashboardElement,
-  isBulkActionElement,
-  isEditableDashboardElement,
-  MultiSelectedEditableDashboardElement,
-} from '../scene/types';
+import { isBulkActionElement } from '../scene/types/BulkActionElement';
+import { EditableDashboardElement, isEditableDashboardElement } from '../scene/types/EditableDashboardElement';
+import { MultiSelectedEditableDashboardElement } from '../scene/types/MultiSelectedEditableDashboardElement';
 
 import { DashboardEditableElement } from './DashboardEditableElement';
 import { MultiSelectedObjectsEditableElement } from './MultiSelectedObjectsEditableElement';
