+++ public/app/features/dashboard-scene/scene/layout-default/DefaultGridLayoutManager.tsx
@@ -22,7 +22,7 @@ import {
   getGridItemKeyForPanelId,
   getDashboardSceneFor,
 } from '../../utils/utils';
-import { DashboardLayoutManager, LayoutRegistryItem } from '../types';
+import { DashboardLayoutManager } from '../types/DashboardLayoutManager';
 
 import { DashboardGridItem } from './DashboardGridItem';
 import { RowRepeaterBehavior } from './RowRepeaterBehavior';
@@ -39,7 +39,16 @@ export class DefaultGridLayoutManager
   extends SceneObjectBase<DefaultGridLayoutManagerState>
   implements DashboardLayoutManager
 {
-  public isDashboardLayoutManager: true = true;
+  public readonly isDashboardLayoutManager = true;
+
+  public static readonly descriptor = {
+    name: 'Default grid',
+    description: 'The default grid layout',
+    id: 'default-grid',
+    createFromLayout: DefaultGridLayoutManager.createFromLayout,
+  };
+
+  public readonly descriptor = DefaultGridLayoutManager.descriptor;
 
   public editModeChanged(isEditing: boolean): void {
     const updateResizeAndDragging = () => {
@@ -328,10 +337,6 @@ export class DefaultGridLayoutManager
     });
   }
 
-  public getDescriptor(): LayoutRegistryItem {
-    return DefaultGridLayoutManager.getDescriptor();
-  }
-
   public cloneLayout(ancestorKey: string, isSource: boolean): DashboardLayoutManager {
     return this.clone({
       grid: this.state.grid.clone({
@@ -400,15 +405,6 @@ export class DefaultGridLayoutManager
         ).children,
       }),
     });
-  }
-
-  public static getDescriptor(): LayoutRegistryItem {
-    return {
-      name: 'Default grid',
-      description: 'The default grid layout',
-      id: 'default-grid',
-      createFromLayout: DefaultGridLayoutManager.createFromLayout,
-    };
   }
 
   /**
