+++ public/app/features/dashboard-scene/scene/layout-rows/RowItemRepeaterBehavior.ts
@@ -13,7 +13,7 @@ import {
 
 import { isClonedKeyOf, getCloneKey } from '../../utils/clone';
 import { getMultiVariableValues } from '../../utils/utils';
-import { DashboardRepeatsProcessedEvent } from '../types';
+import { DashboardRepeatsProcessedEvent } from '../types/DashboardRepeatsProcessedEvent';
 
 import { RowItem } from './RowItem';
 import { RowsLayoutManager } from './RowsLayoutManager';
