+++ public/app/features/dashboard-scene/scene/layouts-shared/layoutRegistry.ts
@@ -3,12 +3,8 @@ import { Registry } from '@grafana/data';
 import { DefaultGridLayoutManager } from '../layout-default/DefaultGridLayoutManager';
 import { ResponsiveGridLayoutManager } from '../layout-responsive-grid/ResponsiveGridLayoutManager';
 import { RowsLayoutManager } from '../layout-rows/RowsLayoutManager';
-import { LayoutRegistryItem } from '../types';
+import { LayoutRegistryItem } from '../types/LayoutRegistryItem';
 
 export const layoutRegistry: Registry<LayoutRegistryItem> = new Registry<LayoutRegistryItem>(() => {
-  return [
-    DefaultGridLayoutManager.getDescriptor(),
-    ResponsiveGridLayoutManager.getDescriptor(),
-    RowsLayoutManager.getDescriptor(),
-  ];
+  return [DefaultGridLayoutManager.descriptor, ResponsiveGridLayoutManager.descriptor, RowsLayoutManager.descriptor];
 });
