+++ public/app/features/dashboard-scene/scene/types/BulkActionElement.ts
@@ -1,0 +1,8 @@
+export interface BulkActionElement {
+  onDelete(): void;
+  onCopy?(): void;
+}
+
+export function isBulkActionElement(obj: object): obj is BulkActionElement {
+  return 'onDelete' in obj;
+}
