+++ public/app/features/dashboard-scene/scene/types/DashboardRepeatsProcessedEvent.ts
@@ -1,0 +1,10 @@
+import { BusEventWithPayload } from '@grafana/data';
+import { SceneObject } from '@grafana/scenes';
+
+export interface DashboardRepeatsProcessedEventPayload {
+  source: SceneObject;
+}
+
+export class DashboardRepeatsProcessedEvent extends BusEventWithPayload<DashboardRepeatsProcessedEventPayload> {
+  public static type = 'dashboard-repeats-processed';
+}
